Add `BagThin` data structure
The `BagThin` data structure is a bag where each element contains only a
`T*`. Unlike `Bag`, where elements can have an additional field,
`ThinBag` is optimised to be space-efficient when this is not necessary.

This change refactors the original bag implementation into a common base
class which is then composed into concrete `Bag` and `BagThin`
implementations.